refactor: rename `command` mod to `process`, move restart_application
