+++ .changes/command-api-module.md
@@ -1,0 +1,5 @@
+---
+"tauri": patch
+---
+
+Renamed the `command` API module to `process`.
