+++ .changes/restart-application.md
@@ -1,0 +1,5 @@
+---
+"tauri": patch
+---
+
+Move `restart_application` API from `app` module to `process` module.
